24.02

+++ ReCapProject/Entities/DTOs/CarDetailDto.cs
@@ -10,7 +10,7 @@ namespace Entities.DTOs
         public int CarId { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
-        public int DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string BrandName { get; set; }
         public string ColorName { get; set; }
     }
